requirements.txt
autopep8==1.5.7
beautifulsoup4==4.9.3
bs4==0.0.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
selenium==3.141.0
soupsieve==2.2.1
toml==0.10.2
urllib3==1.26.6
